{{cookiecutter.dir_name}}/requirements.txt
better_exceptions
aiomysql~=0.1.1
aioredis~=2.0.1
aiosqlite~=0.17.0
anyio~=4.1.0
APScheduler~=3.10.4
async-timeout~=4.0.3
asyncio~=3.4.3
atlastk~=0.13.3
bcrypt~=4.0.0
better-exceptions~=0.3.3
certifi~=2023.11.17
click~=8.1.7
colorama~=0.4.6
DBUtils~=3.0.3
ecdsa~=0.18.0
fastapi~=0.95.0
h11~=0.14.0
httpcore~=0.16.3
httpx~=0.23.3
idna~=3.6
iso8601~=1.1.0
loguru~=0.6.0
passlib~=1.7.4
pyasn1~=0.5.1
pycryptodome~=3.19.0
pydantic~=1.10.13
PyExecJS~=1.5.1
PyJWT~=2.7.0
PyMySQL~=1.0.3
pypika-tortoise~=0.1.6
python-jose~=3.3.0
python-multipart~=0.0.6
pytz~=2023.3.post1
PyYAML~=6.0.1
redis~=2.10.6
rfc3986~=1.5.0
rsa~=4.9
six~=1.16.0
sniffio~=1.3.0
starlette~=0.26.1
tortoise-orm~=0.20.0
typing_extensions~=4.9.0
tzdata~=2023.3
tzlocal~=5.2
uvicorn~=0.20.0
win32-setctime~=1.1.0